requirements.txt
dj-database-url==0.4.2
Django==1.11.4
django-braces==1.11.0
psycopg2==2.7.3
pytz==2017.2
whitenoise==3.3.0
